# cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9044",
    "basePath": "/",
    "paths": {
...
        "/api/user-linked-accounts": {
            "get": {
                "description": "get the users linked accounts.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "root"
                ],
                "summary": "get the users linked accounts.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_user_identity_info.UserLinkedAccounts"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/wellknown_echo.RestErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/wellknown_echo.RestErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/user-linked-accounts/{identity}": {
            "delete": {
                "description": "delete a users linked identity.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "root"
                ],
                "summary": "delete a users linked identity.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "identity name",
                        "name": "identity",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/wellknown_echo.RestErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/wellknown_echo.RestErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api_user_identity_info.Identity": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "api_user_identity_info.UserLinkedAccounts": {
            "type": "object",
            "properties": {
                "identities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api_user_identity_info.Identity"
                    }
                }
            }
        },
...
        "wellknown_echo.RestErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}